requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
pandas>=2.0
openpyxl>=3.1
reportlab>=4.2
django-cors-headers>=4.4
apscheduler>=3.10
PyPDF2>=3.0
# Optionnels:
# pywin32>=306           # notifications Outlook locales (Windows seulement)
# psycopg2-binary>=2.9   # si vous passez à PostgreSQL
